// modules/domain/players/player.entity.d.ts
export declare class Player {
    id: string;
    provider: string;
    provider_id: number;
    name: string;
    name_zh: string;
    position: string;
    nationality: string;
    nationality_code: string;
    birth_date: string;
    height_cm: number;
    weight_kg: number;
    preferred_foot: string;
    photo_url: string;
    meta: Record<string, any>;
    created_at: Date;
    updated_at: Date;
}
//# sourceMappingURL=player.entity.d.ts.map